requirements.txt
amqp
APScheduler
asgiref
attrs
autobahn
Automat
babel
billiard
bleach
celery
certifi
cffi
channels
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
constantly
cron_descriptor
cryptography
daphne
dj-user-login-history
Django
django-auditlog
django-colorfield
django-countries
django-environ
django-filter
django-money
django-multiselectfield
django-summernote
django-timezone-field
djangorestframework
djangorestframework_simplejwt
drf-yasg
et_xmlfile
hyperlink
idna
incremental
inflection
kombu
numpy
oauthlib
openpyxl
packaging
pandas
pillow
prompt_toolkit
py-moneyed
pyasn1
pyasn1_modules
pycountry
pycparser
PyJWT
pyOpenSSL
python-crontab
python-dateutil
pytz
PyYAML
redis
reportlab
requests
requests-oauthlib
service-identity
setuptools
six
sqlparse
Twisted
txaio
typing_extensions
tzdata
tzlocal
uritemplate
urllib3
vine
wcwidth
webencodings
whitenoise
zope.interface
